backend/requirements.txt
flask
flask-cors
tensorflow-cpu
numpy
pillow
scipy